requirements.txt
astor==0.8.1
attrs==25.3.0
babel==2.17.0
bce-python-sdk==0.9.35
blinker==1.8.2
cachetools==5.5.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
config==0.5.1
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
Django==5.1.2
django-request==1.6.3
dotenv==0.9.9
easyocr==1.7.2
einops==0.8.1
et_xmlfile==2.0.0
fastapi==0.118.0
filelock==3.18.0
Flask==3.0.3
flask-babel==4.0.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
imagesize==1.4.1
matplotlib==3.10.3
more-itertools==10.7.0
mpmath==1.3.0
numpy==1.26.4
packaging==24.2
paddleocr==3.0.0
paddlepaddle==3.2.0
paddlex==3.0.0
pandas==1.5.3
pdf2image==1.17.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
PyMuPDF==1.25.5
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
sympy==1.14.0
tiktoken==0.9.0
tokenizers==0.19.1
torch==2.7.0
torchvision==0.22.0
